requirements.txt
st_pages==0.4.3
pyDOE==0.3.8
diversipy==0.8
sklearn==1.3.0
plotly.graph_objects==5.15.0
pyparsing==3.1.1
streamlit-extras==0.3.0
